requirements.txt
arrow==1.2.3
bw==1.5.87
contourpy==1.0.7
cycler==0.11.0
decorator==5.1.1
docopt==0.6.2
Flask==2.2.3
Flask-Cors==3.0.10
itsdangerous==2.1.2
kiwisolver==1.4.4
MarkupSafe==2.1.2
matplotlib==3.7.1
opencv-python==4.7.0.72
pandas==1.5.3
pipreqs==0.4.11
psycopg2-binary==2.9.5
py==1.11.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==1.0.0
retry==0.9.2
ta-bitwarden-cli==0.8.2
Werkzeug==2.2.3
yarg==0.1.9
shapely~=2.0.1
requests~=2.28.2
rasterio~=1.3.6
packaging==23.1
Pillow==9.5.0
pycocotools==2.0.6
pyshp==2.3.1
PyWavelets==1.4.1
scikit-image==0.20.0
scipy==1.10.1
six==1.16.0
sympy==1.12
tifffile==2023.4.12
torch==2.0.1
tqdm==4.65.0
triton==2.0.0
typing_extensions==4.5.0
cmake==3.26.3
filelock==3.12.0
fonttools==4.39.4
imageio==2.28.1
Jinja2==3.1.2
lazy_loader==0.2
lit==16.0.5
mpmath==1.3.0
networkx==3.1
numpy==1.24.3
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
imageio==2.28.1
tifffile==2023.4.12
PyWavelets==1.4.1
scikit_image==0.20.0